+++ app/scripts/services/settings.js
@@ -1219,14 +1219,14 @@ angular.module('accessimapEditeurDerApp')
             'polygon': [
                 { icon: '../bower_components/material-design-icons/action/svg/production/ic_delete_48px.svg', action: deleteService.deleteObject },
                 { icon: '../bower_components/material-design-icons/editor/svg/production/ic_linear_scale_48px.svg', action: moveService.movePoint },
-                { icon: '../bower_components/material-design-icons/image/svg/production/ic_texture_48px.svg', action: 'changePattern' },
-                { icon: '../bower_components/material-design-icons/image/svg/production/ic_color_lens_48px.svg', action: 'changeFillColor' },
+                { icon: '../bower_components/material-design-icons/image/svg/production/ic_texture_48px.svg', action: styleService.changePattern },
+                { icon: '../bower_components/material-design-icons/image/svg/production/ic_color_lens_48px.svg', action: styleService.changeColor },
                 { icon: '../bower_components/material-design-icons/image/svg/production/ic_crop_din_48px.svg', action: styleService.toggleStroke }
             ],
             'circle': [
                 { icon: '../bower_components/material-design-icons/action/svg/production/ic_delete_48px.svg', action: deleteService.deleteObject },
-                { icon: '../bower_components/material-design-icons/image/svg/production/ic_texture_48px.svg', action: 'changePattern' },
-                { icon: '../bower_components/material-design-icons/image/svg/production/ic_color_lens_48px.svg', action: 'changeFillColor' },
+                { icon: '../bower_components/material-design-icons/image/svg/production/ic_texture_48px.svg', action: styleService.changePattern },
+                { icon: '../bower_components/material-design-icons/image/svg/production/ic_color_lens_48px.svg', action: styleService.changeColor },
                 { icon: '../bower_components/material-design-icons/image/svg/production/ic_crop_din_48px.svg', action: styleService.toggleStroke }
             ],
             'text': [