+++ CHANGES.md
@@ -13,6 +13,7 @@ CHANGELOG
 * Rotate a point
 * Improve the line/polygon vertex edition
 * Optionnaly add an arrow at line end
+* Edit color and pattern with radial menu
 
 0.5
 -------------------